Split configs and added init EH
